cositas en api-users

+++ users-api/domain_users/token.go
@@ -2,6 +2,6 @@ package domain_users
 
 type TokenDto struct {
 	Token   string `json:"token"`
-	User_id int    `json:"id_user"`
+	User_id int64  `json:"id_user"`
 	Admin   bool   `json:"admin"`
 }
